Overhauled html, js and style files
In my HTML I overhauled the layout of the page, placing the buttons over
the grid container, the grid size over the grid and changed the footer.

In the style sheet I overhauled all of the styling by adding color
to the page, text-shadows, button shadows and clear borders for each
div on the page.

In the JavaScript file now everything is working as intended per the
project parameters.